+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add metrics instrumentation for sqlalchemy
   ([#1645](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1645))
+- Add connection attributes to sqlalchemy connect span
+  ([#1608](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1608))
 
 ## Version 1.16.0/0.37b0 (2023-02-17)